move grid search to utilities
